move affine atoms from bivariate to affine
