+++ include/affine.h
@@ -41,6 +41,29 @@ expr *new_reshape(expr *child, int d1, int d2);
 expr *new_broadcast(expr *child, int target_d1, int target_d2);
 expr *new_diag_vec(expr *child);
 expr *new_transpose(expr *child);
-expr *new_diag_vec(expr *child);
+
+/* Left matrix multiplication: A @ f(x) where A is a constant sparse
+ * matrix */
+expr *new_left_matmul(expr *u, const CSR_Matrix *A);
+
+/* Left matrix multiplication: A @ f(x) where A is a constant dense
+ * matrix (row-major, m x n). Uses CBLAS for efficient computation. */
+expr *new_left_matmul_dense(expr *u, int m, int n,
+                            const double *data);
+
+/* Right matrix multiplication: f(x) @ A where A is a constant
+ * matrix */
+expr *new_right_matmul(expr *u, const CSR_Matrix *A);
+
+expr *new_right_matmul_dense(expr *u, int m, int n,
+                             const double *data);
+
+/* Constant scalar multiplication: a * f(x) where a is a constant
+ * double */
+expr *new_const_scalar_mult(double a, expr *child);
+
+/* Constant vector elementwise multiplication: a . f(x) where a is
+ * constant */
+expr *new_const_vector_mult(const double *a, expr *child);
 
 #endif /* AFFINE_H */
