+++ src/affine/left_matmul.c
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "bivariate.h"
+#include "affine.h"
 #include "subexpr.h"
 #include "utils/matrix.h"
 #include <assert.h>
