+++ src/bivariate_restricted_dom/quad_over_lin.c
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "bivariate.h"
+#include "bivariate_restricted_dom.h"
 #include "subexpr.h"
 #include "utils/CSC_Matrix.h"
 #include <assert.h>
